+++ lib/kl/primitives/generic_functions.rb
@@ -17,6 +17,13 @@ module Kl
       define_method 'eval-kl' do |exp|
         __eval(exp)
       end
+
+      # Freeze is also provided as a normal function to support
+      # the partial application of zero arguments case. In this
+      # case it becomes a non-special-form.
+      def freeze(exp)
+        Kernel.lambda { exp }
+      end
     end
   end
 end