+++ HISTORY.md
@@ -8,6 +8,7 @@
   - `tlstr` now raises an error when applied to an empty string
   - `set` and `value` now require their first arugment to be a symbol
   - `hd` and `tl` now raise an error when applied to non-lists
+  - `freeze` now supports partial application with zero arguments
 
 ## 0.3.1 - January 24, 2013
 - No code changes. Updates Gemspec to refer to Shen 8.0.